backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.11.0